// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const PostStatus = {
  DRAFT: 'DRAFT',
  SCHEDULED: 'SCHEDULED',
  POSTED: 'POSTED'
} as const

export type PostStatus = (typeof PostStatus)[keyof typeof PostStatus]


export const PostPurpose = {
  ANNOUNCE: 'ANNOUNCE',
  LEARN: 'LEARN',
  DAILY: 'DAILY',
  RECAP: 'RECAP',
  OTHER: 'OTHER'
} as const

export type PostPurpose = (typeof PostPurpose)[keyof typeof PostPurpose]